bashrc: re-add bash_completion stuff
I thought we wouldn't need this, since configuring VS Code to start bash
with "--login" made bash completion work without it. Further reading,
however, made clear that running every terminal as a login shell is not
recommended.
So this is still necessary, even though it is basically enabled system-
wide via /etc/profile.d/bash_completion.sh, which is called from
/etc/profile. The profile files, however, are called only once per
login, so we have to set this again for each bash instance.

See for reference:

https://superuser.com/a/183980/1036029
https://unix.stackexchange.com/a/3085/317320