Add git-fuzzy-diff
Move "git-fuzzy-log.sh" to "git-fzf.sh" and add git-fuzzy-diff function,
which displays a list of all files contained in the diff and a preview
window with the diff of the selected file.
Use this new function within git-fuzzy-log for displaying the diff for
a commit.
Add a git alias for it and make arguments pass through.