git-fuzzy-diff: include rename/move information
If a file was renamed or moved, the according information got lost in
the previous implementation due to how "git diff" works when called with
file arguments (see https://stackoverflow.com/q/71268388/3018229).
Improved the commands now so that the information is visible in the
diff.

Furthermore improved the "enter" action so that the complete file is
shown there instead of only the diff context, since that is already
visible in the preview window.