bash_functions: make order locale-independent
For some reason the sourcing order of the files in bash_functions.d was
not the same on all systems (especially the order of z.sh and zfz.sh,
which is important). The order seems to be affected by the system
locale. Use a pipe to "sort" with an explicitly set locale now in order
to fix this.