forgit: add git-forgit script
This is a part of forgit which I originally didn't notice and hence
ignored for the dotfiles. It is quite useful, though, since it enables
executing forgit functions via a git subcommand "git forgit <function>".
This makes definition of git aliases easier and git will automatically
auto-complete these as soon we have auto-completion for git-forgit.
All the forgit configuration variables have to be exported in order to
be seen by git-forgit.

Took git-forgit from here:
https://github.com/wfxr/forgit/commit/089853173e382be2d41ccbeadd56b361547a1481