forgit: update to 22.12.0
The forgit functionality has been moved to the standalone git-forgit
executable in this release, which makes the bash/zsh plugin obsolete for
me. Hence moved the executable to .local/bin and removed the plugin.

forgit log already uses forgit diff on entering a commit in this
release, so removed the custom FORGIT_LOG_ENTER_COMMAND.