bash_aliases: Fix random "ll" error "bash: COMP_WORDS: bad array subscript"
This commit fixes a seemingly random error which occured sometimes when
calling "ll":

  bash: COMP_WORDS: bad array subscript

The problem was that as soon as one hits tab completion on "ls", exa's
tab completion function is loaded into the environment. This function is
also called "_exa" and overwrites the custom one used in the alias.

Rename the function to prevent this collision.